requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
argparse
python-dotenv
pandas
azure-identity
azure-storage-queue
openai
pydantic
PyMuPDF